Allow exporting multiple certificates with the same local part
Previously, only a single certificate would be exported for a local part.

After this patch is applied WKD Exporter will export all certificates
that have User IDs with the same e-mail address which may be useful
for certificate rotation or delivering code signing certificates
alongside encryption capable ones.

Note that the current implementation does not support merging
certificates. It means that if the input stream contains the same
certificate twice it will also be present in the WKD structure twice.
This is not a problem in practice since the OpenPGP implementation
will perform merging anyway.

The new feature is enabled by the `append` flag that is off by default.

Fixes: https://github.com/wiktor-k/wkd-exporter/issues/2
Signed-off-by: Wiktor Kwapisiewicz <wiktor@metacode.biz>